Add files via upload
A replicable business model for the (i) corporate structure, (ii) data and reporting infrastructure, (iii) risk management and hedging policy, (iv) treasury function and (v) staffing of a company owning and operating containerized mobile data center units. The server capacity is to be used to mine cryptocurrency in this plan.